chart1: Refactor: Split out serial ports ser_ports.cpp/.h library.
The serial port support, mostly enumeration, is a well-defined  set of
about 850 lines. Split this to a separate ser_ports library, trimming
the size of the large chart1.cpp and bringing some more order to the
dependency graph.
